+++ src/wai/annotations/image_utils/__init__.py
@@ -6,3 +6,4 @@ from ._image import image_to_numpyarray
 from ._image import remove_alpha_channel
 from ._masks import mask_to_polygon
 from ._masks import polygon_to_minrect
+from ._masks import polygon_to_lists